+++ ugs-platform/ugs-platform-start-page/src/main/java/com/willwinder/ugp/welcome/FeaturesTab.java
@@ -16,11 +16,11 @@
     You should have received a copy of the GNU General Public License
     along with UGS.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.willwinder.ugp;
+package com.willwinder.ugp.welcome;
 
-import com.willwinder.ugp.content.AbstractTab;
-import com.willwinder.ugp.content.Constants;
-import com.willwinder.ugp.content.ContentSection;
+import com.willwinder.ugp.welcome.content.AbstractTab;
+import com.willwinder.ugp.welcome.content.Constants;
+import com.willwinder.ugp.welcome.content.ContentSection;
 import java.awt.Image;
 import java.awt.Toolkit;
 import java.util.Collection;
