+++ ugs-platform/ugs-platform-start-page/src/main/java/com/willwinder/ugp/welcome/StartPageOptions.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.willwinder.ugp;
+package com.willwinder.ugp.welcome;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
