+++ ugs-platform/ugs-platform-start-page/src/main/resources/com/willwinder/ugp/welcome/Bundle.properties
