+++ ugs-platform/ugs-platform-start-page/src/main/resources/com/willwinder/ugp/welcome/resources/connect@2x.png
