+++ ugs-platform/ugs-platform-start-page/src/main/resources/com/willwinder/ugp/welcome/resources/tab_rollover.png
